Reco: moved matchDirectSimOutputs to a dedicated include file
