+++ SimPPS/DirectSimProducer/python/profile_base_cff.py
@@ -31,20 +31,3 @@ profile_base = cms.PSet(
         efficienciesPerPlane = cms.VPSet()
     )
 )
-
-def matchDirectSimOutputs(process, AOD=False, miniAOD=False):
-    process.totemRPUVPatternFinder.tagRecHit = cms.InputTag('ppsDirectProtonSimulation')
-    process.ctppsPixelLocalTracks.tag = cms.InputTag('ppsDirectProtonSimulation')
-    process.ctppsDiamondLocalTracks.recHitsTag = cms.InputTag('ppsDirectProtonSimulation')
-    if AOD:
-        process.beamDivergenceVtxGenerator.src = cms.InputTag('')
-        process.beamDivergenceVtxGenerator.srcGenParticle = cms.VInputTag(
-            cms.InputTag('genPUProtons', 'genPUProtons'),
-            cms.InputTag('genParticles')
-        )
-    elif miniAOD:
-        process.beamDivergenceVtxGenerator.src = cms.InputTag('')
-        process.beamDivergenceVtxGenerator.srcGenParticle = cms.VInputTag(
-            cms.InputTag('genPUProtons', 'genPUProtons'),
-            cms.InputTag('prunedGenParticles')
-        )
